requirement.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.3
django-celery-results==2.4.0
django-utils-six==2.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
Pillow==9.3.0
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
python-decouple==3.6
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
